Move Signature and related to own module
