+++ src/lib.rs
@@ -17,40 +17,22 @@ mod field;
 mod group;
 mod ecdh;
 mod ecdsa;
-mod ecmult;
+pub mod ecmult;
 mod error;
 mod scalar;
 
-use hmac_drbg::HmacDRBG;
-use sha2::Sha256;
-use typenum::U32;
-
 use scalar::Scalar;
-
-use ecmult::{ECMULT_CONTEXT, ECMULT_GEN_CONTEXT};
 
 pub use error::Error;
 
 pub mod keys;
-pub use keys::{SecretKey, PublicKey};
-
-/// Curve related structs.
-pub mod curve {
-    pub use field::Field;
-    pub use group::{Affine, AffineStorage, Jacobian, AFFINE_G, CURVE_B};
-    pub use scalar::Scalar;
-
-    pub use ecmult::{ECMultContext, ECMultGenContext, ECMULT_CONTEXT, ECMULT_GEN_CONTEXT};
-}
-
+pub mod signature;
 pub mod util;
 
-#[derive(Debug, Clone, Eq, PartialEq)]
-/// An ECDSA signature.
-pub struct Signature {
-    pub r: Scalar,
-    pub s: Scalar,
-}
+pub use keys::{SecretKey, PublicKey};
+pub use signature::{ Signature, verify, recover};
+pub use ecdh::SharedSecret;
+
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
 /// Tag used for public key recovery from signatures.
@@ -60,28 +42,6 @@ pub struct RecoveryId(u8);
 /// Hashed message input to an ECDSA signature.
 pub struct Message(pub Scalar);
 
-#[derive(Debug, Clone, Eq, PartialEq)]
-/// Shared secret using ECDH.
-pub struct SharedSecret([u8; 32]);
-
-impl Signature {
-    pub fn parse(p: &[u8; 64]) -> Signature {
-        let mut r = Scalar::default();
-        let mut s = Scalar::default();
-
-        r.set_b32(array_ref!(p, 0, 32));
-        s.set_b32(array_ref!(p, 32, 32));
-
-        Signature { r, s }
-    }
-
-    pub fn serialize(&self) -> [u8; 64] {
-        let mut ret = [0u8; 64];
-        self.r.fill_b32(array_mut_ref!(ret, 0, 32));
-        self.s.fill_b32(array_mut_ref!(ret, 32, 32));
-        ret
-    }
-}
 
 impl Message {
     pub fn parse(p: &[u8; 32]) -> Message {
@@ -122,62 +82,5 @@ impl Into<i32> for RecoveryId {
     }
 }
 
-impl SharedSecret {
-    pub fn new(pubkey: &PublicKey, seckey: &SecretKey) -> Result<SharedSecret, Error> {
-        let inner = match ECMULT_CONTEXT.ecdh_raw(&pubkey.0, &seckey.0) {
-            Some(val) => val,
-            None => return Err(Error::InvalidSecretKey),
-        };
 
-        Ok(SharedSecret(inner))
-    }
-}
 
-impl AsRef<[u8]> for SharedSecret {
-    fn as_ref(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-/// Check signature is a valid message signed by public key.
-pub fn verify(message: &Message, signature: &Signature, pubkey: &PublicKey) -> bool {
-    ECMULT_CONTEXT.verify_raw(&signature.r, &signature.s, &pubkey.0, &message.0)
-}
-
-/// Recover public key from a signed message.
-pub fn recover(
-    message: &Message,
-    signature: &Signature,
-    recovery_id: &RecoveryId,
-) -> Result<PublicKey, Error> {
-    ECMULT_CONTEXT
-        .recover_raw(&signature.r, &signature.s, recovery_id.0, &message.0)
-        .map(|v| PublicKey(v))
-}
-
-/// Sign a message using the secret key.
-pub fn sign(message: &Message, seckey: &SecretKey) -> Result<(Signature, RecoveryId), Error> {
-    let seckey_b32 = seckey.0.b32();
-    let message_b32 = message.0.b32();
-
-    let mut drbg = HmacDRBG::<Sha256>::new(&seckey_b32, &message_b32, &[]);
-    let generated = drbg.generate::<U32>(None);
-    let mut nonce = Scalar::default();
-    let mut overflow = nonce.set_b32(array_ref!(generated, 0, 32));
-
-    while overflow || nonce.is_zero() {
-        let generated = drbg.generate::<U32>(None);
-        overflow = nonce.set_b32(array_ref!(generated, 0, 32));
-    }
-
-    let result = ECMULT_GEN_CONTEXT.sign_raw(&seckey.0, &message.0, &nonce);
-    #[allow(unused_assignments)]
-        {
-            nonce = Scalar::default();
-        }
-    if let Ok((sigr, sigs, recid)) = result {
-        return Ok((Signature { r: sigr, s: sigs }, RecoveryId(recid)));
-    } else {
-        return Err(result.err().unwrap());
-    }
-}
